FEAT: (tests not passing) add circuit breaker to controllers, two subscribers, and backup interface adapter
Yes, this commit is out of control. I moved directories around and had to touch many places to add the circuit breaker.

Tests are not passing due to a few changes when I added the circuit breaker for AzureBackupInterfaceStoreAdapter. I only included what was needed to get it running for Express+TypeORM.

Remaining problems to resolve are a few small changes to code and a lot of test cleanup for the ABISA/queue circuit breaker.

BUT the Express+Typeorm code runs and the circuit breaker on the queue works.